requirements.txt
pillow
pyautogui
numpy
opencv-python
deskew
scikit-image